Restructure `io` and `boards`
