+++ .github/workflows/ci.yml
@@ -9,8 +9,6 @@ env:
 
 jobs:
   build_and_test:
-    # Temporarily disable until Cargo workspaces / target issues are resolved
-    if: false
     name: Cargo Build & Test
     runs-on: ubuntu-latest
     strategy:
@@ -22,6 +20,7 @@ jobs:
     steps:
       - uses: actions/checkout@v4
       - run: rustup update ${{ matrix.toolchain }} && rustup default ${{ matrix.toolchain }}
+      - run: cargo build --verbose
       - run: rustup target add thumbv7em-none-eabihf
       - run: cargo build --verbose
-      - run: cargo test --verbose
+        working-directory: boards/stm32l476rg
