+++ .github/workflows/lint.yml
@@ -4,14 +4,13 @@ on: push
 
 jobs:
   rust-clippy:
-    # Temporarily disable (see ci.yml)
-    if: false
     name: Run clippy
     runs-on: ubuntu-latest
     steps:
       - uses: actions/checkout@v4
       - name: Run clippy
-        run: cargo clippy --all-targets --all-features -- -D warnings
+        # Add back --all-targets when `error[E0463]: can't find crate for `test`` is fixed
+        run: cargo clippy --all-features -- -D warnings
 
   rust-fmt:
     name: Run rustfmt
