+++ boards/stm32l476rg/Cargo.toml
@@ -1,7 +1,7 @@
 [package]
-edition = "2021"
-name = "stm-i2c-temp-test-board"
+name = "hyped_boards_stm32l476rg"
 version = "0.1.0"
+edition = "2021"
 
 [dependencies]
 embassy-stm32 = { version = "0.1.0", features = ["defmt", "stm32l476rg", "memory-x", "unstable-pac", "time-driver-any", "exti"] , git = "https://github.com/embassy-rs/embassy", rev = "1c466b81e6af6b34b1f706318cc0870a459550b7"}
@@ -23,8 +23,17 @@ critical-section = "1.1"
 embedded-storage = "0.3.1"
 static_cell = "2"
 
-hyped_io_stm32l476rg = { path = "../../lib/io/stm32l476rg" }
+hyped_io = { path = "../../lib/io" }
 hyped_sensors = { path = "../../lib/sensors" }
 
-[profile.release]
-debug = 2
+[lib]
+name = "hyped_boards_stm32l476rg"
+path = "src/lib.rs"
+test = false
+doctest = false
+
+[[bin]]
+name = "test_temp"
+path = "src/bin/temp_test.rs"
+test = false
+doctest = false
