+++ boards/stm32l476rg/src/io.rs
@@ -1,0 +1,1 @@
+pub mod i2c;
