+++ boards/stm32l476rg/src/lib.rs
@@ -1,0 +1,4 @@
+#![no_std]
+
+pub mod io;
+pub mod tasks;
