+++ boards/stm32l476rg/src/tasks.rs
@@ -1,0 +1,1 @@
+pub mod temperature;
