--- lib/io/stm32l476rg/Cargo.toml
@@ -1,9 +1,0 @@
-[package]
-name = "hyped_io_stm32l476rg"
-version = "0.1.0"
-edition = "2021"
-
-[dependencies]
-embassy-stm32 = { version = "0.1.0", features = ["defmt", "stm32l476rg", "memory-x", "unstable-pac", "time-driver-any", "exti"] , git = "https://github.com/embassy-rs/embassy", rev = "1c466b81e6af6b34b1f706318cc0870a459550b7"}
-
-hyped_io = { path = "../" }
